Moved GameServerApp to GameServerConsole
