+++ .gitmodules
@@ -1,4 +1,3 @@
-[submodule "GameServerApp/Content/Data/LeagueSandbox-Default"]
-	path = GameServerApp/Content/Data/LeagueSandbox-Default
+[submodule "LeagueSandbox-Default"]
+	path = LeagueSandbox-Default
 	url = https://github.com/LeagueSandbox/LeagueSandbox-Default.git
-	branch = indev
