+++ GameMode/LeagueSandbox-Default/GameMode.json
