+++ GameServerConsole/Properties/AssemblyInfo.cs
