+++ GameServerConsole/Properties/Resources.resx
