+++ GameServerConsole/Settings/GameServerSettings.json.template
