+++ GameServerConsoleTests/GameServerConsoleTests.csproj
@@ -6,8 +6,8 @@
     <ProjectGuid>{77FE4FF5-3B1C-42A5-B56F-4CAD91013410}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>LeagueSandbox.GameServerAppTests</RootNamespace>
-    <AssemblyName>GameServerAppTests</AssemblyName>
+    <RootNamespace>LeagueSandbox.GameServeConsoleTests</RootNamespace>
+    <AssemblyName>GameServerConsoleTests</AssemblyName>
     <TargetFrameworkVersion>v4.6.2</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <ProjectTypeGuids>{3AC096D0-A1C2-E12C-1390-A8335801FDAB};{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}</ProjectTypeGuids>
@@ -78,9 +78,9 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="..\GameServerApp\GameServerApp.csproj">
+    <ProjectReference Include="..\GameServerConsole\GameServerConsole.csproj">
       <Project>{9ce7ffaa-c3ae-4688-930e-3c71a0cc2f7e}</Project>
-      <Name>GameServerApp</Name>
+      <Name>GameServerConsole</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
