+++ GameServerConsoleTests/Properties/AssemblyInfo.cs
