+++ GameServerLib/Maps/Map.cs
@@ -31,7 +31,6 @@ namespace LeagueSandbox.GameServer.Maps
             Id = _game.Config.GameConfig.Map;
             var path = Path.Combine(
                 game.Config.ContentPath,
-                "Data",
                 _game.Config.ContentManager.GameModeName,
                 "AIMesh",
                 "Map" + Id,
