+++ GameServerLibTests/GameServerLibTests.csproj
@@ -103,9 +103,9 @@
   <Import Project="$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets" Condition="Exists('$(VSToolsPath)\TeamTest\Microsoft.TestTools.targets')" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <_ContentSourceDirectory>$(SolutionDir)GameServerApp\Content</_ContentSourceDirectory>
-    <_ContentTargetDirectory>$(TargetDir)Content\</_ContentTargetDirectory>
-    <_SettingsSourceDirectory>$(SolutionDir)GameServerApp\Settings</_SettingsSourceDirectory>
+    <_ContentSourceDirectory>$(SolutionDir)GameMode</_ContentSourceDirectory>
+    <_ContentTargetDirectory>$(TargetDir)Content\GameMode\</_ContentTargetDirectory>
+    <_SettingsSourceDirectory>$(SolutionDir)GameServerConsole\Settings</_SettingsSourceDirectory>
     <_SettingsTargetDirectory>$(TargetDir)Settings\</_SettingsTargetDirectory>
     <_LibSourceDirectory>$(SolutionDir)lib</_LibSourceDirectory>
     <_LibTargetDirectory>$(TargetDir)</_LibTargetDirectory>
